+++ ListableUI/Sources/Item/Item.swift
@@ -128,6 +128,10 @@ public struct Item<Content:ItemContent> : AnyItem
         self.content
     }
     
+    public var reappliesToVisibleView: ReappliesToVisibleView {
+        self.content.reappliesToVisibleView
+    }
+    
     // MARK: AnyItem_Internal
     
     public func anyIsEquivalent(to other : AnyItem) -> Bool