+++ ListableUI/Sources/ListView/ListView.VisibleContent.swift
@@ -62,29 +62,6 @@ extension ListView
             )
         }
         
-        func updateVisibleViews(with environment : ListEnvironment, animated : Bool)
-        {
-            // Perform Updates Of Visible Headers & Footers
-            
-            self.headerFooters.forEach {
-                $0.headerFooter.applyToVisibleView(with: environment)
-            }
-            
-            // Perform Updates Of Visible Items
-            
-            let update = {
-                self.items.forEach {
-                    $0.item.applyToVisibleCell(with: environment)
-                }
-            }
-            
-            if animated {
-                UIView.animate(withDuration: 0.2, animations: update)
-            } else {
-                update()
-            }
-        }
-        
         private func calculateVisibleContent(in view : ListView) -> (Set<Item>, Set<HeaderFooter>)
         {
             let visibleFrame = view.collectionView.bounds